fix: moved filter functions and rename packages
